Started working on paper

+++ Exercises/Module 1: Numbers/error_propagation_es/error_propagation.c
@@ -3,14 +3,14 @@
 int main() {
 
     // NOTE: la somma e' commutativa per valori in questo range di decimali
-    double x = (0.7 + 0.1) + 0.3;
-    double y = 0.7 + (0.1 + 0.3);
+    float x = (0.7 + 0.1) + 0.3;
+    float y = 0.7 + (0.1 + 0.3);
 
     printf("Left: %.16f, Right: %.16f \n", x, y);
 
-    double xt = 1.e20;
-    double yt = -1.e20;
-    double zt = 1;
+    float xt = 1.e20;
+    float yt = -1.e20;
+    float zt = 1;
 
     // NOTE: la somma non e' piu' commutativa per valori cosi' alti e' necessaria
     // una maggiore precisione (per esempio una quadrupla precisione) per far
